자신의 글만 수정 가능

+++ src/main/java/com/prj2/controller/board/BoardController.java
@@ -4,6 +4,7 @@ import com.prj2.domain.board.Board;
 import com.prj2.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -49,18 +50,31 @@ public class BoardController {
         return ResponseEntity.ok().body(board);
     }
 
-    @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
-        boardService.remove(id);
-    }
-
     @PutMapping("/edit")
+    @PreAuthorize("isAuthenticated()")
     public ResponseEntity edit(@RequestBody Board board,
                                Authentication authentication) {
+        if (!boardService.hasAccess(board.getId(), authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         if (!boardService.validate(board)) {
             return ResponseEntity.badRequest().build();
         }
+        
         boardService.edit(board);
         return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/{id}")
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity delete(@PathVariable Integer id,
+                                 Authentication authentication) {
+
+        if (boardService.hasAccess(id, authentication)) {
+            boardService.remove(id);
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 }
